integrate wand_locker with wand_server
Keeping the locker as a separate client was a nice idea but it had too
much of an overhead to keep everything synchronised. Integrating with
the server seems nicer.

While I was at it, I fixed some logic bugs, cleaned up the server RPC
interface (now in wand.server) and simplified the server code

One functional change: we now scale the lock gain by the poll time