Allow no switch configuration
The wavemeter gets very confused if trying to use it without a switch; this is not a common configuration, but can come up in test environments. This commit allows usage of a config file with a single channel, `"switch": {"type": null}"`  and omission of any channel data in the individual channel config.

This is a bit hacky in that if one were to configure multiple channels and a null switch, there is no way to check this. We rely on user intelligence.